Generar Transcripciones/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
filelock==3.18.0
fsspec==2025.7.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
pillow==11.3.0
proglog==0.1.12
python-dotenv==1.1.1
regex==2025.7.34
requests==2.32.4
sympy==1.14.0
tiktoken==0.10.0
torch==2.8.0+cu129
torchvision==0.23.0+cu129
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
